Move `StringPool` and `StreamLineSpanReader` to `Microsoft.NET.StringTools`
These are general purpose utility classes, suited to this library.
