+++ src/StringTools.UnitTests/StringTools.UnitTests.net35.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <!-- The .NET Core version of MSBuild doesn't support targeting .NET Framework 3.5.  So in that case, we import
        a .props file that prevents building the project from doing much of anything. -->
@@ -15,6 +15,11 @@
     <IsLibraryTestProject>true</IsLibraryTestProject>
     <DefineConstants>$(DefineConstants);NET35_UNITTEST</DefineConstants>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="StreamLineSpanReader_Tests.cs" />
+    <Compile Remove="StringPool_Tests.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Shouldly" />
